+++ src/main/kotlin/icu/windea/pls/model/constants/CwtConfigTextPatternSets.kt
@@ -1,17 +1,9 @@
 package icu.windea.pls.model.constants
 
+import icu.windea.pls.model.constants.CwtConfigTextPatterns as C
+
 object CwtConfigTextPatternSets {
-    val dynamicValueReference = listOf(
-        CwtConfigTextPatterns.value,
-        CwtConfigTextPatterns.valueSet,
-        CwtConfigTextPatterns.dynamicValue,
-    )
-    val singleAliasReference = listOf(
-        CwtConfigTextPatterns.singleAliasRight,
-    )
-    val aliasReference = listOf(
-        CwtConfigTextPatterns.aliasName,
-        CwtConfigTextPatterns.aliasMatchLeft,
-        CwtConfigTextPatterns.aliasKeysField,
-    )
+    val dynamicValueReference = setOf(C.value, C.valueSet, C.dynamicValue)
+    val singleAliasReference = setOf(C.singleAliasRight)
+    val aliasReference = setOf(C.aliasName, C.aliasMatchLeft, C.aliasKeysField)
 }