+++ src/main/kotlin/icu/windea/pls/model/constants/ParadoxDefinitionTypeSets.kt
@@ -1,0 +1,7 @@
+package icu.windea.pls.model.constants
+
+import icu.windea.pls.model.constants.ParadoxDefinitionTypes as C
+
+object ParadoxDefinitionTypeSets {
+    val Job = setOf(C.Job, C.SwappedJob)
+}