+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ## 2.0.2-dev
 
 * [X] #165
+* [X] 优化：优化自动识别切换类型（如`swapped_job`）的声明规则的逻辑 / OP: Optimize the logic for automatically detecting declaration configs for swapped types (e.g. `swapped_job`)
 * [X] 修复：修复未在必要时渲染切分后的图片的问题 / FIX: Fix an issue that the image is not rendered with expected slicing
 * [X] 其他优化与BUG修复 / Other optimizations and bug fixes